using static_cast to cast size variable to 'int '

+++ src/threadpool.cpp
@@ -53,7 +53,7 @@ void threadpool::add_cache_thread()
 
     auto size = tasks_.get_size();
     // (a << b) ==> (a * 2 ^ b)
-    if (size >= (config_.core_threads << 5))
+    if (size >= static_cast<size_t>(config_.core_threads << 5))
     {
         for (int i = config_.core_threads; i < config_.max_threads; i++) 
         {
